AS: Refuse to run differ if an old file is missing.

+++ suite/auto-sync/src/autosync/cpptranslator/Differ.py
@@ -287,6 +287,9 @@ class Differ:
         diff_dir: Path = self.diff_out_dir
         to_copy = self.old_files if self.edit_old_file else self.translated_files
         for f in to_copy:
+            if not f.exists():
+                log.fatal(f"Old version of '{f}' doesn't exist. But the Differ needs an old file for each translated one.")
+                exit(1)
             dest = diff_dir.joinpath(f.name)
             copyfile(f, dest)
             self.patched_file_paths[f.name] = dest
